# Fix critical bug: Keep topic and overall time metrics consistent with per-phase session duration