--- a/C/training/include/progress.h
+++ b/C/training/include/progress.h
@@ -52,24 +52,45 @@ void progress_tracker_destroy(progress_tracker_t *tracker);
 // The num_sessions counter is incremented when progress_tracker_end_session() is called.
 // This is critical for consistency gates to function properly.
 //
+// Time tracking is done per-session, not per-attempt. When a session ends, the session
+// duration is added to:
+// - Phase time metrics (for the current phase)
+// - Topic time metrics (for all topics used during the session)
+// - Overall time metrics
+//
+// This ensures all time metrics grow linearly with actual elapsed time, preventing
+// the quadratic growth bug that occurred when cumulative duration was added per-attempt.
+//
 // Session lifecycle:
 // 1. Call progress_tracker_start_session() to begin a training session
 // 2. Record training attempts using progress_tracker_record_attempt()
-// 3. Call progress_tracker_end_session() to complete the session (increments num_sessions)
+// 3. Call progress_tracker_end_session() to complete the session (increments num_sessions and updates time metrics)
 //
 // Example usage:
 //   progress_tracker_start_session(tracker, PHASE_0_FOUNDATIONS);
 //   for (int i = 0; i < 100; i++) {
 //       bool correct = train_and_check();
-//       progress_tracker_record_attempt(tracker, PHASE_0_FOUNDATIONS, TOPIC_MATH, correct, 1);
+//       progress_tracker_record_attempt(tracker, PHASE_0_FOUNDATIONS, TOPIC_MATH, correct);
 //   }
-//   progress_tracker_end_session(tracker, PHASE_0_FOUNDATIONS);  // num_sessions++
+//   progress_tracker_end_session(tracker, PHASE_0_FOUNDATIONS);  // num_sessions++, time metrics updated
 void progress_tracker_start_session(progress_tracker_t *tracker, uint8_t phase);
 void progress_tracker_end_session(progress_tracker_t *tracker, uint8_t phase);
 bool progress_tracker_is_session_active(progress_tracker_t *tracker, uint8_t phase);
 
-// Record attempts
-void progress_tracker_record_attempt(progress_tracker_t *tracker, curriculum_phase_t phase, topic_type_t topic, bool correct, time_t duration);
+/**
+ * Record a training attempt for a phase and topic.
+ * 
+ * Time is tracked per-session, not per-attempt. Topic and overall time
+ * metrics are updated when the session ends, ensuring consistency with
+ * per-phase time metrics. All time metrics use session duration for
+ * linear growth.
+ * 
+ * @param tracker The progress tracker
+ * @param phase The current phase
+ * @param topic The topic being trained
+ * @param correct Whether the attempt was correct
+ */
+void progress_tracker_record_attempt(progress_tracker_t *tracker, curriculum_phase_t phase, topic_type_t topic, bool correct);
 
 // Query metrics
 void progress_tracker_get_phase_metrics(const progress_tracker_t *tracker, curriculum_phase_t phase, phase_metrics_t *metrics);