## Critical Bug Fix: Time Metrics Consistency

This PR fixes a critical bug where topic and overall time metrics grew quadratically instead of linearly, causing severe inflation (~50x) and inconsistency with per-phase time metrics.

---

## 🐛 Bug Description

**Location:** [`progress_tracker.c:133-149`](https://github.com/Mecca-Research/The-Binary-Decomposition-Interface/blob/990f877d9b56d9fc45f439f7bbe7b803e1291ce2/C/training/curriculum/progress_tracker.c#L133-L149)

After moving `phase_metrics.time_spent` to accumulate once per session (PR #163), `progress_tracker_record_attempt` still added the ever-increasing `duration` argument to `topic_metrics.time_spent` and `overall_metrics.total_time`.

Because `curriculum_submit_answer` supplied `time(NULL) - process->session_start`, those values ballooned **quadratically**: 1+2+3+4+...+n = n(n+1)/2

---

## 💥 Impact

- ❌ **Topic time metrics**: Quadratic growth (~50x inflated for 100 attempts)
- ❌ **Overall time metrics**: Quadratic growth (~50x inflated)
- ✅ **Per-phase time metrics**: Linear growth (correct)
- ❌ **Severe inconsistency** between metric levels
- ❌ **Learning velocity calculations**: Completely wrong
- ❌ **Progress tracking**: Severely inaccurate

### Example of Quadratic Growth

```
Session with 100 attempts over 100 seconds:

Attempt 1 at t=1:  duration=1s  → topic_time += 1s   → total = 1s
Attempt 2 at t=2:  duration=2s  → topic_time += 2s   → total = 3s
Attempt 3 at t=3:  duration=3s  → topic_time += 3s   → total = 6s
...
Attempt 100 at t=100: duration=100s → topic_time += 100s → total = 5,050s

Actual time elapsed: 100 seconds
Recorded topic time: 5,050 seconds (50x inflated!)
Recorded overall time: 5,050 seconds (50x inflated!)
Recorded phase time: 100 seconds (correct!)
```

---

## 🔍 Root Cause

The `duration` parameter passed to `progress_tracker_record_attempt` was **cumulative** (time since session start), not per-attempt. Adding this cumulative value on each attempt caused quadratic growth.

When we fixed session tracking in PR #163, we correctly updated phase metrics to accumulate time once per session. But we forgot to update topic and overall metrics to use the same approach.

---

## ✅ Fix Implementation

### Design Decision: Update topic/overall timing when session ends

**Approach:**
- Remove per-attempt time accumulation from `progress_tracker_record_attempt`
- Track which topics are used during a session
- Update topic and overall time metrics in `progress_tracker_end_session`
- All metrics (phase, topic, overall) use the same notion of elapsed time

**Benefits:**
- ✅ Consistent with phase metric approach
- ✅ Simple implementation
- ✅ No need to track per-attempt timing
- ✅ All metrics use session duration
- ✅ Linear growth for all time metrics

---

## 📝 Changes Made

### 1. **progress_tracker.c**
- Added `topics_used_in_session[TOPIC_MAX]` array to track topics used in current session
- Removed `duration` parameter from `progress_tracker_record_attempt()`
- Removed time accumulation from `record_attempt` (only marks topic as used)
- Updated `progress_tracker_end_session()` to:
  - Add session duration to all topics used in the session
  - Add session duration to overall time
  - Clear topic tracking for next session

### 2. **progress.h**
- Updated function signature (removed `duration` parameter)
- Enhanced documentation explaining session-based time tracking
- Documented that all time metrics use session duration for linear growth

### 3. **curriculum_controller.c**
- Removed `duration` calculation in `curriculum_submit_answer()`
- Removed `duration` parameter when calling `progress_tracker_record_attempt()`

### 4. **test_curriculum.c**
- Updated all test calls to remove `duration` parameter
- Added new `test_time_metrics_consistency()` test that:
  - Records attempts for multiple topics during a session
  - Verifies all time metrics are equal (within 1 second tolerance)
  - Confirms linear growth (not quadratic)

---

## 🧪 Verification

### Test Results

```
=== Running Time Metrics Tests ===

Testing time metrics consistency...
  Expected duration: 2 seconds
  Phase time: 2 seconds
  Math topic time: 2 seconds
  Logic topic time: 2 seconds
  Overall time: 2 seconds
  ✓ All time metrics are consistent (linear growth, not quadratic)
✓ Time metrics consistency test passed

=== All tests passed! ===
```

### Before Fix
```
Session with 100 attempts over 100 seconds:

Phase time:   100s  ✅ (correct - linear growth)
Topic time:   5,050s  ❌ (wrong - quadratic growth, 50x inflated!)
Overall time: 5,050s  ❌ (wrong - quadratic growth, 50x inflated!)

Consistency: ❌ Severe mismatch between metric levels
Learning velocity: ❌ Completely wrong (based on inflated times)
Progress tracking: ❌ Severely inaccurate
```

### After Fix
```
Session with 100 attempts over 100 seconds:

Phase time:   100s  ✅ (correct - linear growth)
Topic time:   100s  ✅ (correct - linear growth)
Overall time: 100s  ✅ (correct - linear growth)

Consistency: ✅ All metrics match
Learning velocity: ✅ Accurate (based on correct times)
Progress tracking: ✅ Accurate
```

---

## 🎯 Success Criteria

- ✅ Topic time metrics grow linearly (not quadratically)
- ✅ Overall time metrics grow linearly (not quadratically)
- ✅ Phase time metrics grow linearly (already correct)
- ✅ All time metrics are consistent (phase = topic = overall for single-topic sessions)
- ✅ Learning velocity calculations are accurate
- ✅ Progress tracking is accurate
- ✅ All tests pass (including new time consistency test)
- ✅ No memory leaks or undefined behavior
- ✅ Documentation updated

---

## 📊 Impact Assessment

This fix is **critical** for accurate progress tracking and learning analytics. Without it:
- All time-based metrics are severely inflated and inconsistent
- Learning velocity calculations are meaningless
- Progress tracking is completely unreliable
- AI process progress is severely understated

With this fix:
- All time metrics are accurate and consistent
- Learning velocity reflects actual learning rate
- Progress tracking is reliable
- Training system analytics are trustworthy

---

## 🔗 Related

- Builds on PR #163 (session tracking bug fix)
- Completes the session-based time tracking implementation
- Ensures consistency across all metric levels